third_party/blink/web_tests/external/wpt/svg/animations/animate-marker-orient-from-auto-to-auto-start-reverse-expected.txt

This is a testharness.js-based test.
[FAIL] Animate SVGMarkerElement orientAttr from auto to auto-start-reverse
  assert_equals: expected (undefined) undefined but got (number) 0
Harness: the test ran to completion.
